Refactor component imports and update package scripts. Changed 'Button' and 'Input' imports to lowercase 'buttom' and 'input' across multiple files. Updated package.json scripts to remove redundant 'rork' command. Cleaned up package-lock.json by removing unused 'react-native-webview' entry.
