+++ app/(tabs)/home.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/Button";
+import Button from "@/components/buttom";
 import { useUser } from "@/hooks/user-store";
 import { router } from "expo-router";
 import { Heart, Sparkles, Users } from "lucide-react-native";
