+++ package.json
@@ -3,9 +3,9 @@
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
-    "start": "bunx rork start -p 0r3d6vmnvu9a2yobhydd6 --tunnel",
-    "start-web": "bunx rork start -p 0r3d6vmnvu9a2yobhydd6 --web --tunnel",
-    "start-web-dev": "DEBUG=expo* bunx rork start -p 0r3d6vmnvu9a2yobhydd6 --web --tunnel",
+    "start": "bunx  start -p 0r3d6vmnvu9a2yobhydd6 --tunnel",
+    "start-web": "bunx  start -p 0r3d6vmnvu9a2yobhydd6 --web --tunnel",
+    "start-web-dev": "DEBUG=expo* bunx  start -p 0r3d6vmnvu9a2yobhydd6 --web --tunnel",
     "lint": "expo lint"
   },
   "dependencies": {
